refactor(api): ddd: migrate oidcAuthenticationServiceRegistry singleton from lib
Migrate the oidcAuthenticationServiceRegistry singleton from
lib/domain/usecases/index.js→api/src/identity to
access-management/domain/usecases/index.js
