+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.admin.controller.js
@@ -1,6 +1,5 @@
-import { oidcAuthenticationServiceRegistry } from '../../../../lib/domain/usecases/index.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
-import { usecases } from '../../domain/usecases/index.js';
+import { oidcAuthenticationServiceRegistry, usecases } from '../../domain/usecases/index.js';
 import * as oidcProviderSerializer from '../../infrastructure/serializers/jsonapi/oidc-identity-providers.serializer.js';
 import { getForwardedOrigin, RequestedApplication } from '../../infrastructure/utils/network.js';
